refactor: Move add-model functionality from a modal to a seperate page.
