+++ src/features/HandleModel/HandleModelComponent/HandleModelComponent.styled.ts
@@ -1,0 +1,14 @@
+import styled from 'styled-components';
+import { spacings } from '../../../tokens/spacings';
+
+export const Wrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+  row-gap: ${spacings.X_LARGE};
+`;
+
+export const CustomContent = styled.div`
+  display: flex;
+  flex-direction: column;
+  row-gap: ${spacings.X_LARGE};
+`;
