+++ src/features/HandleModel/HandleModelComponent/HandleModelComponent.tsx
@@ -1,17 +1,19 @@
 /* eslint-disable max-lines-per-function */
-import { Button, Typography } from '@equinor/eds-core-react';
+import { Button, LinearProgress, Typography } from '@equinor/eds-core-react';
 import { useEffect, useState } from 'react';
 import { AnalogueModelDetail } from '../../../api/generated';
 import { ModelInputFilesTable } from '../ModelInputFilesTable/ModelInputFilesTable';
 import { ModelMetadata } from '../ModelMetadata/ModelMetadata';
-import { useAddModelDialog, validateValues } from './AddModelDialog.hooks';
-import * as Styled from './AddModelDialog.styled';
+import {
+  useHandleModelComponent,
+  validateValues,
+} from './HandleModelComponent.hooks';
+import * as Styled from './HandleModelComponent.styled';
 
 interface AddModelDialogProps {
-  isOpen: boolean;
   confirm?: (file: File, metadata: AnalogueModelDetail) => Promise<void>;
   edit?: (metadata: AnalogueModelDetail) => Promise<void>;
-  cancel: () => void;
+  progress?: number;
   uploading?: boolean;
   defaultMetadata: AnalogueModelDetail;
   isEdit?: boolean;
@@ -38,11 +40,11 @@ const defaultFiles = {
   INI: undefined,
 };
 
-export const AddModelDialog = ({
-  isOpen,
+export const HandleModelComponent = ({
   confirm,
   edit,
-  cancel,
+  progress,
+  // cancel,
   uploading,
   defaultMetadata,
   isEdit,
@@ -57,18 +59,17 @@ export const AddModelDialog = ({
   const [fileSize, setFileSize] = useState(0);
   const [rawFile, setrawFile] = useState<File>();
 
-  useAddModelDialog(setFileSize, setMetadata, files, rawFile, existingData);
+  useHandleModelComponent(
+    setFileSize,
+    setMetadata,
+    files,
+    rawFile,
+    existingData,
+  );
 
   const handleSubmit = () => {
     setErrors(validateValues(metadata, files, isEdit));
     setSubmitting(true);
-  };
-
-  const handleCancle = () => {
-    setMetadata(defaultMetadata);
-    setFiles(defaultFiles);
-    setErrors({});
-    cancel();
   };
 
   const fileAdded = (e: React.ChangeEvent<HTMLInputElement>) => {
@@ -117,13 +118,11 @@ export const AddModelDialog = ({
   const INIFileContent = () => <p>Not implemented yet...</p>;
 
   return (
-    <Styled.Dialog open={isOpen}>
-      <Styled.Dialog.Header>
-        <Styled.Dialog.Title>
-          {isEdit ? 'Edit model details' : 'Add new model'}
-        </Styled.Dialog.Title>
-      </Styled.Dialog.Header>
-      <Styled.DialogCustomContent scrollable>
+    <Styled.Wrapper>
+      <Typography variant="h3">
+        {isEdit ? 'Edit model details' : 'Add new model'}
+      </Typography>
+      <Styled.CustomContent>
         {!isEdit && (
           <ModelInputFilesTable
             files={files}
@@ -145,23 +144,17 @@ export const AddModelDialog = ({
         {Object.keys(errors).includes('file') && (
           <Typography className="error">NC file missing</Typography>
         )}
-      </Styled.DialogCustomContent>
-      <Styled.DialogActions>
-        {uploading && (
-          <Typography>
-            You have to wait until current upload has finnished before a new one
-            can start.{' '}
-          </Typography>
-        )}
-        <div>
-          <Button onClick={handleSubmit} disabled={uploading}>
-            {!isEdit ? 'Confirm and start uploading' : 'Save changes'}
-          </Button>
-          <Button variant="outlined" onClick={handleCancle}>
-            Cancel
-          </Button>
-        </div>
-      </Styled.DialogActions>
-    </Styled.Dialog>
+      </Styled.CustomContent>
+      <div>
+        <Button onClick={handleSubmit} disabled={uploading}>
+          {isEdit
+            ? 'Save changes'
+            : uploading
+            ? 'Wait for model to finish uploading'
+            : 'Confirm and start uploading'}
+        </Button>
+        {uploading && <LinearProgress variant="determinate" value={progress} />}
+      </div>
+    </Styled.Wrapper>
   );
 };
